requirements.txt
MouseInfo==0.1.3
numpy==2.4.2
opencv-python==4.13.0.92
pandas==3.0.0
pillow==12.1.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python3-xlib==0.15
pytweening==1.2.0
six==1.17.0

